+++ test/e2e/pages/Main.page.js
@@ -1,0 +1,9 @@
+export default class MainPage {
+  constructor() {
+    this.title = 'My Page';
+  }
+
+  open(path) {
+    browser.url(path);
+  }
+}
