--- test/pages/Login.page.js
@@ -1,19 +1,0 @@
-import mainPage from './Main.page';
-import { createPublicKey } from 'crypto';
-
-class LoginPage extends mainPage {
-    get loginBtn() { return $('a.button'); }
-
-
-    open() {
-        super.open()
-    }
-
-    clickLoginBtn() {
-        this.loginBtn.click();
-    }
-}
-
-export default new LoginPage();
-// module.exports.LoginPage;
-// module.exports.clickLoginBtn();
