--- test/pages/Main.page.js
@@ -1,9 +1,0 @@
-export default class MainPage {
-    constructor() {
-        this.title = 'My Page'
-    }
-
-    open(path) {
-        browser.url(path)
-    }
-}
